Move all osu!catch legacy skin piece files to the correct location
